app/requirements.txt
fastapi
uvicorn[standard]
qdrant-client
sentence-transformers
httpx
python-dotenv
redis
prometheus-fastapi-instrumentator